Documented methods and classes that were marked with TODO, and reorganised the file hierarchy.
