+++ app/src/main/java/com/africanbongo/whipitkotlin/network/service/RequestStatus.kt
@@ -1,4 +1,4 @@
-package com.africanbongo.whipitkotlin.network
+package com.africanbongo.whipitkotlin.network.service
 
 enum class RequestStatus {
     LOADING, DONE, ERROR
